+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## v1.0.33
 * Fixed fly wheel instance doesn't render with embeddium
 * Added searching feature for the biome / dimension condition
+* Support the APIs of the customizable machine name